requirements.txt
streamlit
pylast
python-dotenv